refactor: move cache to root
